# Conflicts:
#	log/estatemanagement_log.log
#	log/estatemanagement_log.log.2024-08-17.0.gz
#	log/estatemanagement_log.log.2024-08-21.0.gz
#	src/main/java/estate/management/com/EstateManagementApplication.java
#	src/main/java/estate/management/com/domain/user/UserRole.java
#	src/main/java/estate/management/com/payload/request/abstracts/AbstractUserRequest.java
#	src/main/java/estate/management/com/payload/request/abstracts/BaseUserRequest.java
#	src/main/java/estate/management/com/payload/response/ResponseMessage.java
#	src/main/java/estate/management/com/payload/response/abstracts/BaseUserResponse.java
#	src/main/java/estate/management/com/repository/UserRepository.java
#	src/main/java/estate/management/com/repository/business/AdvertRepository.java
#	src/main/java/estate/management/com/repository/business/AdvertTypeRepository.java
#	src/main/java/estate/management/com/repository/business/FavoriteRepository.java
#	src/main/java/estate/management/com/security/config/WebSecurityConfig.java
#	src/main/java/estate/management/com/security/jwt/JwtUtils.java
#	src/main/java/estate/management/com/security/service/UserDetailsImpl.java
#	src/main/java/estate/management/com/service/business/FavoriteService.java
#	src/main/resources/application.properties

+++ src/main/java/estate/management/com/domain/advert/Advert.java
@@ -2,6 +2,7 @@ package estate.management.com.domain.advert;
 
 import com.fasterxml.jackson.annotation.JsonFormat;
 import com.fasterxml.jackson.annotation.JsonIgnore;
+import estate.management.com.domain.category.Category;
 import lombok.*;
 
 import javax.persistence.*;
@@ -110,7 +111,7 @@ public class Advert {
 
     @Column(name = "category_id", nullable = false)
     @NotNull(message = "categoryId cannot be null" )
-    private int categoryId;
+    private Long categoryId;
 
 
     // ------ METHODS --------
